requirements.txt
fastapi>=0.111,<0.116
uvicorn[standard]>=0.29,<0.31
diff-match-patch
langchain-openai>=0.1.20,<0.2.0
pydantic>=2.6,<3.0
python-dotenv>=1.0.1,<2.0
numpy>=1.26,<3.0